refactor: criação de decoder e configuration

+++ access/src/main/java/br/com/vinicius/access/service/AccessService.java
@@ -3,6 +3,8 @@ package br.com.vinicius.access.service;
 import br.com.vinicius.access.client.ClientClient;
 import br.com.vinicius.access.client.DoorClient;
 import br.com.vinicius.access.entity.AccessEntity;
+import br.com.vinicius.access.exceptions.AccessAlreadyExistException;
+import br.com.vinicius.access.exceptions.AccessNotFoundException;
 import br.com.vinicius.access.mapper.AccessMapper;
 import br.com.vinicius.access.model.AccessModel;
 import br.com.vinicius.access.model.ClientModel;
@@ -12,6 +14,8 @@ import org.springframework.data.domain.Example;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class AccessService {
@@ -29,28 +33,38 @@ public class AccessService {
     }
 
     public AccessModel registryAccess(AccessEntity entity){
-        verifyClientAndDoor(entity);
+        if(verifyIfAccessExistAndClientAndDoor(entity)){
+            throw new AccessAlreadyExistException();
+        }
         AccessEntity save = repository.save(entity);
         return mapper.to(save);
     }
 
     @Transactional
     public void deleteAccess(AccessEntity entity){
-        verifyClientAndDoor(entity);
+        if(!verifyIfAccessExistAndClientAndDoor(entity)){
+            throw new AccessNotFoundException();
+        }
         repository.removeByDoorIdAndClientId(entity.getDoorId(), entity.getClientId());
     }
 
     public AccessModel getAccess(AccessEntity entity){
-        verifyClientAndDoor(entity);
-        boolean exists = repository.exists(Example.of(entity));
-        if(exists){
+        if(verifyIfAccessExistAndClientAndDoor(entity)){
             return mapper.to(entity);
         }
-        return null;
+        throw new AccessNotFoundException();
     }
 
-    private void verifyClientAndDoor(AccessEntity entity) {
+    public List<AccessModel> getListAccess() {
+        List<AccessEntity> all = repository.findAll();
+        if(all.isEmpty())
+            throw new AccessNotFoundException();
+        return all.stream().map(entity -> mapper.to(entity)).collect(Collectors.toList());
+    }
+
+    private boolean verifyIfAccessExistAndClientAndDoor(AccessEntity entity) {
         ClientModel client = clientClient.getClient(entity.getClientId());
         DoorModel door = doorClient.getDoor(entity.getDoorId());
+        return repository.exists(Example.of(entity));
     }
 }
